+++ project/src/components/app/app.tsx
@@ -1,6 +1,6 @@
 import { NotFound } from '../not-found';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import MainPage from '../main-page';
+import MainPage from '../main-page/main-page';
 import SignIn from '../sign-in';
 import Favorites from '../favorites/favorites';
 import Room from '../room/room';
